Don't overwrite existing configs

+++ lib/tasks/dev.rake
@@ -20,7 +20,11 @@ namespace :dev do
     config_file = File.join('config', "#{example_base}.yml")
     config_example = File.join('config', "#{example_base}.yml.example")
     t = Rake::FileTask.define_task(config_file => config_example) do |task|
-      cp task.prerequisites.first, task.name
+      if File.exists?(task.name)
+        puts ">>> Skipping #{task.prerequisites.first} -> #{task.name}, please examine and merge if needed."
+      else
+        cp task.prerequisites.first, task.name
+      end
     end
     file_task_names << t.name
   end
